refactor: move model structs to internal/model package
