+++ internal/testutil/fixture/testdata_wiki.go
@@ -1,16 +1,16 @@
 package fixture
 
 import (
-	backlog "github.com/nattokin/go-backlog"
+	"github.com/nattokin/go-backlog/internal/model"
 )
 
 type wikiFixtures struct {
 	MaximumJSON string
-	Maximum     backlog.Wiki
+	Maximum     model.Wiki
 	MinimumJSON string
-	Minimum     backlog.Wiki
+	Minimum     model.Wiki
 	ListJSON    string
-	List        []*backlog.Wiki
+	List        []*model.Wiki
 }
 
 // Wiki provides test fixtures for Wiki-related tests.
@@ -122,67 +122,67 @@ var Wiki = wikiFixtures{
     "updated": "2012-07-23T06:09:48Z"
 }
 `,
-	Maximum: backlog.Wiki{
+	Maximum: model.Wiki{
 		ID:        34,
 		ProjectID: 56,
 		Name:      "Maximum Wiki Page",
 		Content:   "This is a muximal wiki page.",
-		Tags: []*backlog.Tag{
+		Tags: []*model.Tag{
 			{ID: 12, Name: "proceedings"},
 		},
-		Attachments: []*backlog.Attachment{
+		Attachments: []*model.Attachment{
 			{
 				ID:   23,
 				Name: "test.json",
 				Size: 8857,
-				CreatedUser: &backlog.User{
+				CreatedUser: &model.User{
 					ID:          1,
 					UserID:      "admin",
 					Name:        "admin",
-					RoleType:    backlog.RoleAdministrator,
+					RoleType:    model.RoleAdministrator,
 					Lang:        "ja",
 					MailAddress: "eguchi@nulab.example",
 				},
 				Created: mustTime("2014-01-06T11:10:45Z"),
 			},
 		},
-		SharedFiles: []*backlog.SharedFile{
+		SharedFiles: []*model.SharedFile{
 			{
 				ID:   454403,
 				Type: "file",
 				Dir:  "/icon/",
 				Name: "01_buz.png",
 				Size: 2735,
-				CreatedUser: &backlog.User{
+				CreatedUser: &model.User{
 					ID:          5686,
 					UserID:      "takada",
 					Name:        "takada",
-					RoleType:    backlog.RoleNormalUser,
+					RoleType:    model.RoleNormalUser,
 					Lang:        "ja",
 					MailAddress: "takada@nulab.example",
 				},
 				Created: mustTime("2009-02-27T03:26:15Z"),
-				UpdatedUser: &backlog.User{
+				UpdatedUser: &model.User{
 					ID:          5686,
 					UserID:      "takada",
 					Name:        "takada",
-					RoleType:    backlog.RoleNormalUser,
+					RoleType:    model.RoleNormalUser,
 					Lang:        "ja",
 					MailAddress: "takada@nulab.example",
 				},
 				Updated: mustTime("2009-03-03T16:57:47Z"),
 			},
 		},
-		Stars: []*backlog.Star{
+		Stars: []*model.Star{
 			{
 				ID:    75,
 				URL:   "https://xx.backlogtool.com/view/BLG-1",
 				Title: "[BLG-1] first issue | Show issue - Backlog",
-				Presenter: &backlog.User{
+				Presenter: &model.User{
 					ID:          1,
 					UserID:      "admin",
 					Name:        "admin",
-					RoleType:    backlog.RoleAdministrator,
+					RoleType:    model.RoleAdministrator,
 					Lang:        "ja",
 					MailAddress: "eguchi@nulab.example",
 				},
@@ -193,31 +193,31 @@ var Wiki = wikiFixtures{
 				Comment: "ok",
 				URL:     "https://xx.backlogtool.com/view/BLG-1",
 				Title:   "[BLG-1] first issue | Show issue - Backlog",
-				Presenter: &backlog.User{
+				Presenter: &model.User{
 					ID:          1,
 					UserID:      "admin",
 					Name:        "admin",
-					RoleType:    backlog.RoleAdministrator,
+					RoleType:    model.RoleAdministrator,
 					Lang:        "ja",
 					MailAddress: "eguchi@nulab.example",
 				},
 				Created: mustTime("2014-01-23T10:55:19Z"),
 			},
 		},
-		CreatedUser: &backlog.User{
+		CreatedUser: &model.User{
 			ID:          1,
 			UserID:      "admin",
 			Name:        "admin",
-			RoleType:    backlog.RoleAdministrator,
+			RoleType:    model.RoleAdministrator,
 			Lang:        "ja",
 			MailAddress: "eguchi@nulab.example",
 		},
 		Created: mustTime("2012-07-23T06:09:48Z"),
-		UpdatedUser: &backlog.User{
+		UpdatedUser: &model.User{
 			ID:          1,
 			UserID:      "admin",
 			Name:        "admin",
-			RoleType:    backlog.RoleAdministrator,
+			RoleType:    model.RoleAdministrator,
 			Lang:        "ja",
 			MailAddress: "eguchi@nulab.example",
 		},
@@ -258,31 +258,31 @@ var Wiki = wikiFixtures{
     "updated": "2012-07-23T06:09:48Z"
 }
 `,
-	Minimum: backlog.Wiki{
+	Minimum: model.Wiki{
 		ID:        34,
 		ProjectID: 56,
 		Name:      "Minimum Wiki Page",
 		Content:   "This is a minimal wiki page.",
-		Tags: []*backlog.Tag{
+		Tags: []*model.Tag{
 			{ID: 12, Name: "proceedings"},
 		},
-		Attachments: []*backlog.Attachment{},
-		SharedFiles: []*backlog.SharedFile{},
-		Stars:       []*backlog.Star{},
-		CreatedUser: &backlog.User{
+		Attachments: []*model.Attachment{},
+		SharedFiles: []*model.SharedFile{},
+		Stars:       []*model.Star{},
+		CreatedUser: &model.User{
 			ID:          1,
 			UserID:      "admin",
 			Name:        "admin",
-			RoleType:    backlog.RoleAdministrator,
+			RoleType:    model.RoleAdministrator,
 			Lang:        "ja",
 			MailAddress: "eguchi@nulab.example",
 		},
 		Created: mustTime("2012-07-23T06:09:48Z"),
-		UpdatedUser: &backlog.User{
+		UpdatedUser: &model.User{
 			ID:          1,
 			UserID:      "admin",
 			Name:        "admin",
-			RoleType:    backlog.RoleAdministrator,
+			RoleType:    model.RoleAdministrator,
 			Lang:        "ja",
 			MailAddress: "eguchi@nulab.example",
 		},
@@ -350,26 +350,26 @@ var Wiki = wikiFixtures{
     }
 ]
 `,
-	List: []*backlog.Wiki{
+	List: []*model.Wiki{
 		{
 			ID:        112,
 			ProjectID: 56,
 			Name:      "test1",
-			Tags:      []*backlog.Tag{{ID: 12, Name: "proceedings"}},
-			CreatedUser: &backlog.User{
+			Tags:      []*model.Tag{{ID: 12, Name: "proceedings"}},
+			CreatedUser: &model.User{
 				ID:          1,
 				UserID:      "admin",
 				Name:        "admin",
-				RoleType:    backlog.RoleAdministrator,
+				RoleType:    model.RoleAdministrator,
 				Lang:        "ja",
 				MailAddress: "eguchi@nulab.example",
 			},
 			Created: mustTime("2013-05-30T09:11:36Z"),
-			UpdatedUser: &backlog.User{
+			UpdatedUser: &model.User{
 				ID:          1,
 				UserID:      "admin",
 				Name:        "admin",
-				RoleType:    backlog.RoleAdministrator,
+				RoleType:    model.RoleAdministrator,
 				Lang:        "ja",
 				MailAddress: "eguchi@nulab.example",
 			},
@@ -379,21 +379,21 @@ var Wiki = wikiFixtures{
 			ID:        115,
 			ProjectID: 56,
 			Name:      "test2",
-			Tags:      []*backlog.Tag{{ID: 12, Name: "proceedings"}},
-			CreatedUser: &backlog.User{
+			Tags:      []*model.Tag{{ID: 12, Name: "proceedings"}},
+			CreatedUser: &model.User{
 				ID:          1,
 				UserID:      "admin",
 				Name:        "admin",
-				RoleType:    backlog.RoleAdministrator,
+				RoleType:    model.RoleAdministrator,
 				Lang:        "ja",
 				MailAddress: "eguchi@nulab.example",
 			},
 			Created: mustTime("2013-05-30T09:11:36Z"),
-			UpdatedUser: &backlog.User{
+			UpdatedUser: &model.User{
 				ID:          1,
 				UserID:      "admin",
 				Name:        "admin",
-				RoleType:    backlog.RoleAdministrator,
+				RoleType:    model.RoleAdministrator,
 				Lang:        "ja",
 				MailAddress: "eguchi@nulab.example",
 			},
